New. Handle basic authorization under student groups.

+++ app/Http/Controllers/Admin/Students/GroupsController.php
@@ -9,18 +9,34 @@ use App\Models\StudentGroup;
 
 class GroupsController extends AdminController
 {
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showAll()
     {
+        $this->authorize('view-groups');
+
         return view('pages.admin.student-groups-list', [
             'groups' => StudentGroup::withCount('students')->get()
         ]);
     }
 
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showNewGroupForm()
     {
+        $this->authorize('create-groups');
+
         return view('pages.admin.student-groups-new');
     }
 
+    /**
+     * @param CreateGroupRequest $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function newGroup(CreateGroupRequest $request)
     {
         $validated = $request->validated();
@@ -31,8 +47,14 @@ class GroupsController extends AdminController
         ]);
     }
 
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showSingleGroup()
     {
+        $this->authorize('view-students');
+
         $group = $this->urlManager->getCurrentGroup();
 
         return view('pages.admin.student-groups-single', [
@@ -40,13 +62,23 @@ class GroupsController extends AdminController
         ]);
     }
 
+    /**
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function showUpdateGroupForm()
     {
+        $this->authorize('update-groups');
+
         return view('pages.admin.student-group-settings', [
             'group' => $this->urlManager->getCurrentGroup()
         ]);
     }
 
+    /**
+     * @param UpdateGroupRequest $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
     public function updateGroup(UpdateGroupRequest $request)
     {
         $group = $this->urlManager->getCurrentGroup();
@@ -57,8 +89,15 @@ class GroupsController extends AdminController
         ]);
     }
 
+    /**
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     * @throws \Exception
+     */
     public function deleteGroup()
     {
+        $this->authorize('delete-groups');
+
         $group = $this->urlManager->getCurrentGroup();
         $group->delete();
